Add CSS, SCSS and HTML files, update example file and JS
Added main.css and main.scss for styling, and index.html for the structure of the page. Modified the example-dxdiag.txt file to add more system information. Also updated main.js file to add a comment for Node.js environment.

These changes provide structure and styling for the web application and update the diagnostic example to cover more system information. The comment was added in main.js to indicate that it's applicable for a Node.js environment.